+++ R/zzz.R
@@ -4,6 +4,6 @@ utils::globalVariables(
         "radius", "angle", "CI_upper", "CI_lower",
         "CI_lower_X", "CI_lower_Y", "CI_upper_X", "CI_upper_Y",
         "index", "X_new", "Y_new", "CI_X", "CI_Y", "se.fit",
-        ".index", "chn"
+        ".index", "chn", "coord", "coord_knot", "value"
     )
 )